api/users/views.py
from rest_framework import viewsets, mixins

from users.models import User
from users.serializers import UserSerializer


class UserCreateViewSet(mixins.CreateModelMixin,
                        viewsets.GenericViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
