requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2==2.9.1
drf-yasg==1.20.0
